refactor: split index.ts into state/ and ui/ modules
Extract three self-contained concerns out of the 1.5kloc index.ts mount():

- state/filterState.ts (119 lines): VALID_KINDS, DEFAULT_KINDS, STORAGE_KEY,
  loadFilterState, saveFilterState, computeVisibleNodeIds. Pure functions
  with no closure dependencies.

- state/physicsSnapshot.ts (144 lines): localStorage persistence with a
  createPhysicsSnapshotIO() factory that owns the throttle state. Caller
  provides graphUrl, simNodes, and a camera-state getter explicitly.

- ui/Overlays.ts (124 lines): createLoadingOverlay, createChainOverlay,
  createOnboardingOverlay. The container element (and theme for the chain
  chip) become explicit parameters so the factories don't have to live
  inside mount()'s closure.

index.ts: 1573 -> 1278 lines. Pure restructuring; no behavior change.
typecheck, tests (10/10), and format:check all green.
